refactor: choice step and step with attempt classes
